pyproject.toml
[project]
name = "news-buddy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.125.0",
    "trafilatura>=2.0.0",
    "feedparser>=6.0.0",
    "httpx>=0.28.0",
    "selectolax>=0.3.0",
    "apscheduler>=3.10.0",
    "google-generativeai>=0.8.0",
    "jinja2>=3.1.0",
    "uvicorn>=0.34.0",
]
